Split index into submodules
